Straightforward breaking out code from Create.
Most notably, factored out `ActionSupport::Hasher` from new
`UsersController::Action::Update::Internals::UserDataFilter' and
existing `UsersController::Action::Create::Internals::PostDataFilter`
classes.

RSpec: 732 examples, 0 failures
RuboCop: 249 files inspected, no offenses detected
